New material handling with gadts
New material handling with GADTs seems to work well.
I use it for overloading constructors. But they are
good for passing around constraints.
The constant medium implementation has been started.
The triangle between scatter material and hitrecord
needs to be resolved somehow. It is very ugly to
implement materials this way.