First batch of refactoring done
First batch of refactoring is done. Most objects
that work with lists now require a non empty list.
The functions that require random behaviour needs
to be truly refactored. The current makeshift design
reflects the desire for what is to be done but
does not do it so well.
Ray tracing can be thought of a sequence of random
requiring steps with different transformations occuring
in between. So we need to bind/sticth those steps
together to make the sequence and need to traverse
that sequence and to be able to apply functions
to purer aspect for obtaining transformations.
This is what haskell is conceived for.
So I should be able to pull it off.