Spectral rendering started
Spectral rendering started. There shall be 2 options.
We either specify the light as a planar wave both in P and S
polarizations, then compute the reflected and refracted
planar waves in same polarizations using the material
properties. The Fresnel module that is supposed to do that
compiles.
The other option is that we use sampled spectral distributions.
Where the light is expressed as power over different wavelengths.
We can specify this spectral distribution as either a function
or a series of values.
Then we create a spectral material and spectral texture that fits
to color computation of both of these options