SampledSpectrum is underway
Sampled spectrum is underway.
pbr_spectrum methods needs to be converted.
then sampled spectrum methods will be converted.
It should not be too difficult to make the passage as
the pbr spectrum methods use a lot of lambda functions.

A good part of the sampled spectrum is concerned with
converting spectrums to RGB values. The idea is to ignore
the generated wavelength of the camera ray for materials
with RGB textures and only deal with objects with spectral
materials and spectral textures. Spectral textures output
power values upon receiving wavelengths, we then convert them
to RGB values when we combine them with other texture types.