Started color interface
Now the general architecture is that the
texture should output spectrum.
A spectrum can be an rgb model or a sampled power
values at given wavelength range.
In order to construct a colorable type that can
be either an rgb model or a sampled power
we need to be able to construct sampled power
from rgb values and we need to be able to output
rgb values from sampled power. The same thing
goes for rgb model, we need to be able to construct
a sampled spectrum using values provided by pbrt.