The spectral handling is almost done
The spectral handling of colors are almost done.
The idea is the following:
Each rayColor produces a ColorInterface
each mkColor produces PixelSpectrum.
We decide what to produce by looking at the background.
If background color interface is rgb model, the evaluation
does not consider wavelength of the ray.
If the background color interface is some other model
the evaluation considers wavelength as well.
Since the fresnel equation do not change wavelength,
we should be okay