Nothing compiles started type change.
Two functions require a type change for waves.
Init of PBR spectrum requires some cie lambda to be real.
I almost always prefered word for waves which turned out
to be an unwise decision.
Now have to fix that code so that it compiles.
Then I can fill the pbr spectrum functions etc.