Spectral image
Spectral output produces an image. However, it does not
produce correct colors. Everything is produced as
either blue with higher values leading to pinkish color.
Or just they are not visible. We are probably encoding colors
wrong in fromRGB or toXYZ. Not sure which one. The scaling also
eliminates colors, I am not sure why.