Here is the next batch of things to do
Currently the spectral texture accepts a sampled spectrum.
The goal is to extend it so that it would accept a full spectrum,
that is a function that generates power values given some
wavelength.
This can be done as the following: The plane wave produce
an electric field. An electric field is a vector field
with units. The energy, is the magnitude of the field.
It is the length of the vector.
The plane wave should be parametrized with amplitudes
and phase angles.
In our case amplitudes are related using fresnel equations.
Once we have the wave, its wavelength is dependent on
its wavenumber. The wavenumber is angular frequency
divided by velocity of the wave. In the case of
light waves, the velocity of the wave is the speed
of light about 3*10^8, for other waves like sound wave
the velocity changes.
The wavelength is the 2 * π divided by the wavenumber.
So the angular frequency determines the wavenumber which
in turn determines the wavelength.
The angular frequency also determines the electric field
magnitude due to wave equation.
The full spectrum texture then is simply a texture that
outputs the magnitude of the electric field given a wavelength.
The incoming light is contains a wavelength which can be
transformed into angular phase. The amplitude of the outgoing
ray can be computed using fresnel properties of the surface.
With amplitude and the angular frequency of the outgoing ray
in place, we should be able to compute the power value,
that is the magnitude of the electric field